+++ vulcanai2/models/__init__.py
@@ -1,5 +1,10 @@
 # -*- coding: utf-8 -*-
 """ Imports network classes so they can be used directly"""
+from .basenetwork import BaseNetwork
+from .cnn import ConvNet
+from .dnn import DenseNet
+from .ensemble import SnapshotNet
+
 __all__ = [
     'basenetwork',
     'cnn',
@@ -7,11 +12,9 @@ __all__ = [
     'layers',
     'ensemble',
     'metrics',
-    'utils'
+    'utils',
+    'BaseNetwork',
+    'ConvNet',
+    'DenseNet',
+    'SnapshotNet'
 ]
-
-# Commenting this leads to circular imports when
-# running test on plotters
-# from .cnn import ConvNet
-# from .dnn import DenseNet
-# from .ensemble import SnapshotNet
