+++ vulcanai2/plotters/__init__.py
@@ -1,5 +1,22 @@
 # -*- coding: utf-8 -*-
+from .visualization import (
+    compute_saliency_map,
+    display_saliency_overlay,
+    display_pca,
+    display_tsne,
+    display_confusion_matrix,
+    display_record,
+    display_receptive_fields
+)
+
 __all__ = [
     'utils',
-    'visualization'
+    'visualization',
+    'compute_saliency_map',
+    'display_saliency_overlay',
+    'display_pca',
+    'display_tsne',
+    'display_confusion_matrix',
+    'display_record',
+    'display_receptive_fields'
 ]
